requirements.txt
Flask==3.0.3
pandas==2.2.2
requests==2.32.3
google-cloud-bigquery==3.25.0
pyarrow==17.0.0
db-dtypes==1.3.0
gunicorn==22.0.0